Move regex AST tests into test-regex-ast.
