+++ tests/test-lexer.cpp
@@ -52,22 +52,6 @@ using log_surgeon::SchemaVarAST;
 
 namespace {
 /**
- * Generates an AST for the given `var_schema` string. Then serialize the AST and compare it with
- * the `expected_serialize_ast`.
- * @param var_schema
- * @param expected_serialized_ast
- */
-auto test_regex_ast(string_view var_schema, u32string const& expected_serialized_ast) -> void;
-
-/**
- * Converts the characters in a 32-byte unicode string into 4 bytes to generate a 8-byte unicode
- * string.
- * @param u32_str
- * @return The resulting utf8 string.
- */
-[[nodiscard]] auto u32string_to_string(u32string const& u32_str) -> string;
-
-/**
  * Lexes the given input and verifies the output is a token for the given rule name, folowed by the
  * end of input token.
  *
@@ -91,26 +75,6 @@ auto test_scanning_input(
  * @return The serialized map.
  */
 [[nodiscard]] auto serialize_id_symbol_map(unordered_map<rule_id_t, string> const& map) -> string;
-
-auto test_regex_ast(string_view const var_schema, u32string const& expected_serialized_ast)
-        -> void {
-    Schema schema;
-    schema.add_variable(var_schema, -1);
-
-    auto const schema_ast = schema.release_schema_ast_ptr();
-    auto const* capture_rule_ast
-            = dynamic_cast<SchemaVarAST*>(schema_ast->m_schema_vars.at(0).get());
-    REQUIRE(capture_rule_ast != nullptr);
-
-    auto const actual_string = u32string_to_string(capture_rule_ast->m_regex_ptr->serialize());
-    auto const expected_string = u32string_to_string(expected_serialized_ast);
-    REQUIRE(actual_string == expected_string);
-}
-
-auto u32string_to_string(u32string const& u32_str) -> string {
-    wstring_convert<codecvt_utf8<char32_t>, char32_t> converter;
-    return converter.to_bytes(u32_str.data(), u32_str.data() + u32_str.size());
-}
 
 auto test_scanning_input(
         LogParser& log_parser,
@@ -212,7 +176,9 @@ TEST_CASE("Test the Schema class", "[Schema]") {
         auto& schema_var_ast = dynamic_cast<SchemaVarAST&>(*schema_var_ast_ptr);
         REQUIRE(var_name == schema_var_ast.m_name);
 
-        REQUIRE_NOTHROW([&]() { dynamic_cast<RegexASTCatByte&>(*schema_var_ast.m_regex_ptr); }());
+        REQUIRE_NOTHROW([&]() {
+            (void)dynamic_cast<RegexASTCatByte&>(*schema_var_ast.m_regex_ptr);
+        }());
     }
 
     SECTION("Add a capture variable to schema") {
@@ -261,79 +227,6 @@ TEST_CASE("Test the Schema class", "[Schema]") {
         REQUIRE(1 == regex_ast_group_ast->get_ranges().size());
         REQUIRE('0' == regex_ast_group_ast->get_ranges().at(0).first);
         REQUIRE('9' == regex_ast_group_ast->get_ranges().at(0).second);
-    }
-
-    SECTION("Test AST with tags") {
-        // This test validates the serialization of a regex AST with named capture groups. The
-        // serialized output includes tags (<n> for positive matches, <~n> for negative matches) to
-        // indicate which capture groups are matched or unmatched at each node.
-        test_regex_ast(
-                // clang-format off
-                "capture:"
-                "Z|("
-                    "A(?<letter>("
-                        "(?<letter1>(a)|(b))|"
-                        "(?<letter2>(c)|(d))"
-                    "))B("
-                        "?<containerID>\\d+"
-                    ")C"
-                ")",
-                U"(Z<~letter1><~letter2><~letter><~containerID>)|("
-                    "A("
-                        "(((a)|(b))<letter1><~letter2>)|"
-                        "(((c)|(d))<letter2><~letter1>)"
-                    ")<letter>B("
-                        "([0-9]){1,inf}"
-                    ")<containerID>C"
-                ")"
-                // clang-format on
-        );
-    }
-
-    SECTION("Test repetition regex") {
-        // Repetition without capture groups untagged and tagged AST are the same
-        test_regex_ast("capture:a{0,10}", U"()|((a){1,10})");
-        test_regex_ast("capture:a{5,10}", U"(a){5,10}");
-        test_regex_ast("capture:a*", U"()|((a){1,inf})");
-        test_regex_ast("capture:a+", U"(a){1,inf}");
-
-        // Repetition with capture groups untagged and tagged AST are different
-        test_regex_ast("capture:(?<letter>a){0,10}", U"(<~letter>)|(((a)<letter>){1,10})");
-        test_regex_ast("capture:(?<letter>a){5,10}", U"((a)<letter>){5,10}");
-        test_regex_ast("capture:(?<letter>a)*", U"(<~letter>)|(((a)<letter>){1,inf})");
-        test_regex_ast("capture:(?<letter>a)+", U"((a)<letter>){1,inf}");
-
-        // Capture group with repetition
-        test_regex_ast("capture:(?<letter>a{0,10})", U"(()|((a){1,10}))<letter>");
-
-        // Complex repetition
-        test_regex_ast(
-                // clang-format off
-                "capture:"
-                "("
-                    "("
-                        "(?<letterA>a)|"
-                        "(?<letterB>b)"
-                    ")*"
-                ")|("
-                    "("
-                        "(?<letterC>c)|"
-                        "(?<letterD>d)"
-                    "){0,10}"
-                ")",
-                U"("
-                    U"(<~letterA><~letterB>)|(("
-                        U"((a)<letterA><~letterB>)|"
-                        U"((b)<letterB><~letterA>)"
-                    U"){1,inf})"
-                U"<~letterC><~letterD>)|("
-                    U"(<~letterC><~letterD>)|(("
-                        U"((c)<letterC><~letterD>)|"
-                        U"((d)<letterD><~letterC>)"
-                    U"){1,10})"
-                U"<~letterA><~letterB>)"
-                // clang-format on
-        );
     }
 }
